restructuring file layout
