+++ src/analysis/system_properties.rs
@@ -1,281 +1,15 @@
 extern crate libc;
 extern crate netlib_src; // enable linking with blas and lapack
-use crate::{Continuous, Discrete, ss::Ss, tf::Tf, traits::Time};
-use libc::{c_char, c_double, c_int};
+// use crate::{Continuous, Discrete, ss::Ss, tf::Tf, traits::Time};
+use crate::{
+    slicot_wrapper::{ab08nd_, ab13bd_, ab13dd_, dggev_},
+    systems::Ss,
+    utils::traits::{Continuous, Discrete, Time},
+};
+use libc::{c_double, c_int};
 use nalgebra::DMatrix;
 use num_complex::{Complex64, c64};
 use std::{error::Error, ffi::CString};
-
-unsafe extern "C" {
-    /// SLICOT transform state-space to transfer matrix. See [SLICOT documentation](https://github.com/TorBorve/SLICOT-Reference)
-    fn tb04ad_(
-        rowcol: *const c_char,
-        n: *const c_int,
-        m: *const c_int,
-        p: *const c_int,
-        a: *mut c_double,
-        lda: *const c_int,
-        b: *mut c_double,
-        ldb: *const c_int,
-        c: *mut c_double,
-        ldc: *const c_int,
-        d: *const c_double,
-        ldd: *const c_int,
-        nr: *mut c_int,
-        index: *mut c_int,
-        dcoeff: *mut c_double,
-        lddcoe: *const c_int,
-        ucoeff: *mut c_double,
-        lduco1: *const c_int,
-        lduco2: *const c_int,
-        tol1: *const c_double,
-        tol2: *const c_double,
-        iwork: *mut c_int,
-        dwork: *mut c_double,
-        ldwork: *const c_int,
-        info: *mut c_int,
-    );
-
-    // Calculate H2 or L2 norm of State space system
-    fn ab13bd_(
-        dico: *const c_char,
-        jobn: *const c_char,
-        n: *const c_int,
-        m: *const c_int,
-        p: *const c_int,
-        a: *mut c_double,
-        lda: *const c_int,
-        b: *mut c_double,
-        ldb: *const c_int,
-        c: *mut c_double,
-        ldc: *const c_int,
-        d: *mut c_double,
-        ldd: *const c_int,
-        nq: *mut c_int,
-        tol: *const c_double,
-        dwork: *mut c_double,
-        ldwork: *const c_int,
-        iwarn: *mut c_int,
-        info: *mut c_int,
-    ) -> c_double;
-
-    // Calculate H-inf norm of State space system
-    fn ab13dd_(
-        dico: *const c_char,
-        jobe: *const c_char,
-        equil: *const c_char,
-        jobd: *const c_char,
-        n: *const c_int,
-        m: *const c_int,
-        p: *const c_int,
-        fpeak: *mut c_double,
-        a: *mut c_double,
-        lda: *const c_int,
-        e: *mut c_double,
-        lde: *const c_int,
-        b: *mut c_double,
-        ldb: *const c_int,
-        c: *mut c_double,
-        ldc: *const c_int,
-        d: *mut c_double,
-        ldd: *const c_int,
-        gpeak: *mut c_double,
-        tol: *const c_double,
-        iwork: *mut c_int,
-        dwork: *mut c_double,
-        ldwork: *const c_int,
-        cwork: *mut c_double,
-        lcwork: *const c_int,
-        info: *mut c_int,
-    );
-
-    // Create regular pencil for system, which can be used to find invariant
-    // zeros
-    fn ab08nd_(
-        equil: *const c_char,
-        n: *const c_int,
-        m: *const c_int,
-        p: *const c_int,
-        a: *const c_double,
-        lda: *const c_int,
-        b: *const c_double,
-        ldb: *const c_int,
-        c: *const c_double,
-        ldc: *const c_int,
-        d: *const c_double,
-        ldd: *const c_int,
-        nu: *mut c_int,
-        rank: *mut c_int,
-        dinfz: *mut c_int,
-        nkror: *mut c_int,
-        nkrol: *mut c_int,
-        infz: *mut c_int,
-        kronr: *mut c_int,
-        kronl: *mut c_int,
-        a_f: *mut c_double,
-        ldaf: *const c_int,
-        b_f: *mut c_double,
-        ldbf: *const c_int,
-        tol: *const c_double,
-        iwork: *mut c_int,
-        dwork: *mut c_double,
-        ldwork: *const c_int,
-        info: *mut c_int,
-    );
-}
-
-// LAPACK function
-unsafe extern "C" {
-    // Compute eigenvalues and vectors for generalized eigenvalue problem
-    fn dggev_(
-        jobvl: *const c_char,
-        jobvr: *const c_char,
-        n: *const c_int,
-        a: *mut c_double,
-        lda: *const c_int,
-        b: *mut c_double,
-        ldb: *const c_int,
-        alphar: *mut c_double,
-        alphai: *mut c_double,
-        beta: *mut c_double,
-        vl: *mut c_double,
-        ldvl: *const c_int,
-        vr: *mut c_double,
-        ldvr: *const c_int,
-        work: *mut c_double,
-        lwork: *const c_int,
-        info: *mut c_int,
-    );
-}
-
-/// Converts a state-space system (A, B, C, D) into a transfer function using
-/// the SLICOT `tb04ad_` function.
-///
-/// # Parameters:
-/// - `a`: A `DMatrix` representing the state matrix `A`.
-/// - `b`: A `DMatrix` representing the input matrix `B`.
-/// - `c`: A `DMatrix` representing the output matrix `C`.
-/// - `d`: A `DMatrix` representing the direct transmission matrix `D`.
-///
-/// # Returns:
-/// A `Result` containing the transfer function (`Tf<f64, U>`) or an error
-/// (`Box<dyn Error>`).
-pub fn ss2tf_tb04ad<U: Time + 'static>(
-    a: &DMatrix<f64>,
-    b: &DMatrix<f64>,
-    c: &DMatrix<f64>,
-    d: &DMatrix<f64>,
-) -> Result<Tf<f64, U>, Box<dyn Error + 'static>> {
-    Ss::<U>::verify_dimensions(a, b, c, d)?;
-
-    let rowcol = CString::new("R")?;
-
-    let n = a.nrows();
-    let m = b.ncols();
-    let p = c.nrows();
-
-    if m != 1 || p != 1 {
-        return Err(Box::new(std::io::Error::new(
-            std::io::ErrorKind::InvalidInput,
-            "State Space system must be SISO single-input-single-output",
-        )));
-    }
-
-    let lda = n;
-    let ldb = n;
-    let ldc = p;
-    let ldd = p;
-
-    let mut a_in = DMatrix::zeros(lda, n);
-    a_in.view_mut((0, 0), (n, n)).copy_from(a);
-    let mut b_in = DMatrix::zeros(ldb, m);
-    b_in.view_mut((0, 0), (n, m)).copy_from(b);
-    let mut c_in = DMatrix::zeros(ldc, n);
-    c_in.view_mut((0, 0), (p, n)).copy_from(c);
-    let mut d_in = DMatrix::zeros(ldd, m);
-    d_in.view_mut((0, 0), (p, m)).copy_from(d);
-
-    let mut nr: c_int = 0;
-    let mut index = vec![0 as c_int; p];
-
-    let lddcoe = p;
-    let mut dcoeff = DMatrix::<f64>::zeros(lddcoe, n + 1);
-
-    let lduco1 = p;
-    let lduco2 = m;
-    assert!(p == 1 && m == 1);
-    let mut ucoeff = DMatrix::<f64>::zeros(lduco1, n + 1);
-
-    // use default tolerances
-    let tol1 = -1. as c_double;
-    let tol2 = -1 as c_double;
-
-    use std::cmp::max;
-    let mut iwork = vec![0 as c_int; n + max(m, p)];
-
-    let mp = m;
-    let pm = p;
-    let ldwork = 10
-        * max(
-            1,
-            n * (n + 1) + max(max(n * mp + 2 * n + max(n, mp), 3 * mp), pm),
-        );
-    let mut dwork = vec![0. as c_double; ldwork];
-    let mut info = 0 as c_int;
-
-    unsafe {
-        tb04ad_(
-            rowcol.as_ptr(),
-            &(n as c_int),
-            &(m as c_int),
-            &(p as c_int),
-            a_in.as_mut_ptr(),
-            &(lda as c_int),
-            b_in.as_mut_ptr(),
-            &(ldb as c_int),
-            c_in.as_mut_ptr(),
-            &(ldc as c_int),
-            d_in.as_mut_ptr(),
-            &(ldd as c_int),
-            &mut nr,
-            index.as_mut_ptr(),
-            dcoeff.as_mut_ptr(),
-            &(lddcoe as c_int),
-            ucoeff.as_mut_ptr(),
-            &(lduco1 as c_int),
-            &(lduco2 as c_int),
-            &tol1,
-            &tol2,
-            iwork.as_mut_ptr(),
-            dwork.as_mut_ptr(),
-            &(ldwork as c_int),
-            &mut info,
-        );
-    }
-
-    if info != 0 {
-        return Err(Box::new(std::io::Error::other(format!(
-            "SLICOT tb04ad_ failed with info code {}",
-            info
-        ))));
-    }
-
-    let den_degree = index[0] as usize;
-    let den: Vec<f64> = dcoeff
-        .view((0, 0), (1, den_degree + 1))
-        .iter()
-        .rev()
-        .copied()
-        .collect();
-    let num: Vec<f64> = ucoeff
-        .view((0, 0), (1, den_degree + 1))
-        .iter()
-        .rev()
-        .copied()
-        .collect();
-    let tf = Tf::<f64, U>::new(num.as_slice(), den.as_slice());
-    Ok(tf)
-}
 
 /// Computes the H2 norm of the system.
 ///
